requirements/default.txt
alembic==0.9.2
Flask==0.12
Flask-API==0.7.1
flask-marshmallow==0.8.0
Flask-Migrate==2.0.4
requests==2.11.1
SQLAlchemy==1.1.13
tweepy==3.5.0
jsonschema==2.6.0
marshmallow-sqlalchemy==0.13.1
selenium==3.4.3
beautifulsoup4==4.6.0
pyperclip==1.5.27
urllib3
python-dateutil==2.6.1
pytz==2017.2
psycopg2==2.6.2
git+https://github.com/dvska/gdata-python3#egg=gdata
